+++ packages/rnv/src/platformTools/firefox.js
@@ -75,7 +75,7 @@ const copyKaiOSAssets = (c, platform) =>
         logTask('copyKaiOSAssets');
         if (!isPlatformActive(c, platform, resolve)) return;
 
-        const sourcePath = path.join(c.appConfigFolder, 'assets', platform);
+        const sourcePath = path.join(c.paths.appConfigFolder, 'assets', platform);
         const destPath = path.join(getAppFolder(c, platform));
 
         copyFolderContentsRecursiveSync(sourcePath, destPath);
