+++ packages/rnv/src/platformTools/tizen.js
@@ -61,7 +61,7 @@ const copyTizenAssets = (c, platform) => new Promise((resolve, reject) => {
     logTask('copyTizenAssets');
     if (!isPlatformActive(c, platform, resolve)) return;
 
-    const sourcePath = path.join(c.appConfigFolder, 'assets', platform);
+    const sourcePath = path.join(c.paths.appConfigFolder, 'assets', platform);
     const destPath = path.join(getAppFolder(c, platform));
 
     copyFolderContentsRecursiveSync(sourcePath, destPath);
